ml-service/requirements.txt
fastapi==0.109.1
pydantic==2.4.0
spacy==3.7.2
scikit-learn==1.3.0
numpy==1.26.0
rake-nltk==1.0.6
keybert==0.7.0
openai>=1.3.0
python-dotenv
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
